nuh uh

+++ src/main/java/json/jayson/client/data/ClientSoulsData.java
@@ -1,25 +1,25 @@
 package json.jayson.client.data;
 
 public class ClientSoulsData {
-    private static int playerSouls;
-    private static int newPlayerSouls;
+    private static long playerSouls;
+    private static long newPlayerSouls;
 
-    private static int newPlayerSoulsTick;
-    public static void set(int souls, int newPlayerSouls, int newPlayerSoulsTick) {
+    private static long newPlayerSoulsTick;
+    public static void set(long souls, long newPlayerSouls, long newPlayerSoulsTick) {
         ClientSoulsData.playerSouls = souls;
         ClientSoulsData.newPlayerSouls = newPlayerSouls;
         ClientSoulsData.newPlayerSoulsTick = newPlayerSoulsTick;
     }
 
-    public static int getNewPlayerSoulsTick() {
+    public static long getNewPlayerSoulsTick() {
         return newPlayerSoulsTick;
     }
 
-    public static int getSouls() {
+    public static long getSouls() {
         return playerSouls;
     }
 
-    public static int getNewPlayerSouls() {
+    public static long getNewPlayerSouls() {
         return newPlayerSouls;
     }
 }
